build: add ErrorProne and NullAway, and fix what they found
Wired on JDK 17 through 23. ErrorProne compiles against javac internals and so
trails new JDK releases; on an unsupported one it fails the compile outright
rather than degrading, which is why it sits behind a version-gated profile. The
JDK 25 job builds the same code without it, leaving that build unanalysed rather
than unbuilt. It also needs javac's internals opened to the Maven JVM, hence
.mvn/jvm.config, and -XDaddTypeAnnotationsToSymbol from JDK 21 onward.

It earned the setup immediately by breaking the build on a real finding, and
four more behind it:

Two OpenTelemetry context scopes were flagged as leaked. They are closed, by
SpanScope, but the analyser cannot follow the hand-off and a try-with-resources
at the reporting site would close the scope before the message was published.
Both are annotated with that reasoning rather than silently ignored.

Two Future values were being discarded in the in-memory transport, and these
mattered. An exception escaping the dispatch loop would have been captured in a
future nobody reads: the dispatcher would stop, the queue would stop being
consumed, and a test would hang rather than fail. The same applied to queue
expiry, where a failure would have stranded a message in a retry rung with
nothing to explain it. Dispatch now uses execute so the exception reaches the
thread's handler, and the expiry task reports its own failures.

Two Javadoc comments had been left stacked on the same method by an earlier edit,
so the documentation that was meant to be read was being discarded. Also removed
dead fields and made an implicit long-to-double conversion explicit.

MissingSummary is switched off: eighty-two reports, all of them accessors
documented by @return alone, which is deliberate here and not a defect.

NullAway stays on as a warning. Its fifty findings are almost all one shape:
builder fields are null until build() is called, which it cannot know without
@Nullable on them. Adding those means adding JSpecify, and acemq-amqp-api is
documented as having no third-party dependencies, so that is a decision to take
deliberately rather than a tidy-up to slip in here. Provided scope would keep the
promise for consumers; it is recorded as the next step.

Verified on JDK 21 with the analysis active and on JDK 25 with the profile
inactive; both build green with coverage at 85.0 %.

+++ acemq-amqp-api/src/main/java/org/acemq/amqp/api/RetryPolicy.java
@@ -123,9 +123,9 @@ public final class RetryPolicy {
         }
 
         List<Duration> schedule = new ArrayList<>();
-        double current = initialDelay.toMillis();
+        double current = (double) initialDelay.toMillis();
         for (int i = 1; i < maxAttempts; i++) {
-            long millis = (long) Math.min(current, maxDelay.toMillis());
+            long millis = (long) Math.min(current, (double) maxDelay.toMillis());
             schedule.add(Duration.ofMillis(millis));
             current *= multiplier;
         }
